Add files via upload
This is code I wrote in Java that checks if user input is a palindrome.
It uses a custom stack built with linked lists to reverse and compare characters. I wrote everything from the interface to the logic myself, focusing on clean structure, input handling, and core data structure implementation.